+++ runtime/darwinia/src/pallets/identity.rs
@@ -5,9 +5,9 @@ use crate::{weights::pallet_identity::WeightInfo, *};
 
 frame_support::parameter_types! {
 	// Minimum 100 bytes/CRING deposited (1 MILLI/byte)
-	pub const BasicDeposit: Balance = 10 * COIN;       // 258 bytes on-chain
-	pub const FieldDeposit: Balance = 250 * MILLI;     // 66 bytes on-chain
-	pub const SubAccountDeposit: Balance = 2 * COIN;   // 53 bytes on-chain
+	pub const BasicDeposit: Balance = deposit(1, 258);
+	pub const FieldDeposit: Balance = deposit(0, 66);
+	pub const SubAccountDeposit: Balance = deposit(1, 53);
 	pub const MaxSubAccounts: u32 = 100;
 	pub const MaxAdditionalFields: u32 = 100;
 	pub const MaxRegistrars: u32 = 20;
