+++ runtime/darwinia/src/pallets/relay_authorities.rs
@@ -10,14 +10,14 @@ use darwinia_relay_primitives::relay_authorities::OpCode;
 
 frame_support::parameter_types! {
 	pub const EthereumRelayAuthoritiesLockId: LockIdentifier = *b"ethrauth";
-	pub const EthereumRelayAuthoritiesTermDuration: BlockNumber = 30 * DAYS;
+	pub const EthereumRelayAuthoritiesTermDuration: BlockNumber = 7 * DAYS;
 	pub const MaxCandidates: usize = 7;
 	pub const OpCodes: (OpCode, OpCode) = (
 		[71, 159, 189, 249],
 		[180, 188, 244, 151]
 	);
 	pub const SignThreshold: Perbill = Perbill::from_percent(60);
-	pub const SubmitDuration: BlockNumber = 100;
+	pub const SubmitDuration: BlockNumber = 300;
 }
 impl Config<EthereumRelayAuthoritiesInstance> for Runtime {
 	type Event = Event;
